Core/Quests: Fixed remaining issues with quest crediting (note: previous commits did not introduce issues, only revealed them) Core/DBLayer: Fixed more compile warnings Core/Spells: Corrected code style from revision 634af79146
Closes issue #3956.

--HG--
branch : trunk

+++ src/server/shared/Database/QueryResult.cpp
@@ -105,7 +105,7 @@ enum Field::DataTypes ResultSet::ConvertNativeType(enum_field_types mysqlType) c
     }
 }
 
-void ResultBind::BindResult(uint32& num_rows)
+void ResultBind::BindResult(uint64& num_rows)
 {
     FreeBindBuffer();
